+++ escodoo_customization/static/src/customization_service.js
@@ -21,7 +21,7 @@ export const customizationService = {
                 if (state.enabled) {
                     notification.add(
                         _t(
-                            "Customization mode is on. Click a field, page, group, button, list column, kanban card field or button, search field or menu."
+                            "Customization mode is on. Click a field, page, group, button, list column, kanban card field or button, kanban header button, progressbar, search field or menu."
                         ),
                         {type: "info"}
                     );